requirements.txt
google-adk
google-auth
google-cloud-bigquery
google-genai
